Change rating data type to numeric

+++ src/main/resources/db/h2/schema.sql
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS education_center (
 	name VARCHAR,
 	city_id INTEGER REFERENCES city(id),
 	registration_fee VARCHAR,
-	rating REAL
+	rating NUMERIC(2, 1)
 );
 
 CREATE TABLE IF NOT EXISTS course (
